+++ docs/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## unreleased
+
+- C4PlantUMLExporter: `c4plantuml.stdlib` is now `true` by default.
+
 ## 1.13.0 (17th March 2023)
 
 - Adds support for groups on dynamic views (Show group also in dynamic view #1).